Removed further warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@20304 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: d186def9534266b9468bcea59b06313753a6948c

+++ server/wrapper/ivef/de.uniol.inf.is.odysseus.wrapper.ivef/src/de/uniol/inf/is/odysseus/wrapper/ivef/element/version_0_1_5/Sensor.java
@@ -16,8 +16,8 @@ import de.uniol.inf.is.odysseus.wrapper.ivef.IIvefElement;
 
 public class Sensor implements IIvefElement { 
 
-    private double m_senId = (float)0; // default value is uninitialized
-    private double m_trkId = (float)0; // default value is uninitialized
+    private double m_senId = 0; // default value is uninitialized
+    private double m_trkId = 0; // default value is uninitialized
 
     public Sensor() {
